Move targetFundTooltip body
