+++ src/steps/transferFutureCapital/targetFundTooltipBody/advanced.png
