+++ src/steps/transferFutureCapital/targetFundTooltipBody/conservative.png
